requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
datasets==4.0.0
decord==0.6.0
deepspeed==0.17.5
dill==0.3.8
einops==0.8.1
filelock==3.19.1
flash-attn==2.6.3
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
h5py==3.14.0
hf-xet==1.1.9
hjson==3.1.0
hnswlib==0.8.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
nltk==3.9.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
pip==25.1
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pynvml==12.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.8
ray==2.49.0
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rpds-py==0.26.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.35.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
tokenizers==0.21.4
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.3
triton==3.1.0
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
